feat(turbomcp): add TurboMcpClientHandler for auto-updating tools and full rmcp parity
- Extract inline turbomcp module to tool/turbomcp.rs with module docs
- Add TurboMcpClientHandler<T> for ToolServer auto-registration with
  notifications/tools/list_changed support via ToolRefreshHandler
- Add TurboMcpClientError with ConnectionError, ToolFetchError,
  ToolServerError variants preserving original error types
- Fix resource content rendering to include MIME type and content body
- Add partial-registration rollback in connect() on failure
- Expand example to full lifecycle: in-process server + ToolServer +
  TurboMcpClientHandler auto-updating pattern
- Add 4 integration tests using channel transport (real protocol path)
- Add 2 new resource rendering tests (with/without MIME type)
- Bump turbomcp dependency to 3.0.11
- Add [patch.crates-io] for local development
