+++ rig/rig-core/Cargo.toml
@@ -41,7 +41,7 @@ thiserror = { workspace = true }
 tracing = { workspace = true }
 url = { workspace = true }
 rmcp = { version = "0.16", optional = true, features = ["client"] }
-turbomcp-client = { version = "3.0.0", optional = true, default-features = false, features = ["http"] }
+turbomcp-client = { version = "3.0.11", optional = true, default-features = false, features = ["http"] }
 tokio = { workspace = true, features = ["rt", "sync"] }
 http = "1.3.1"
 tracing-futures = { version = "0.2.5", features = ["futures-03"] }
@@ -98,11 +98,11 @@ rmcp = { version = "0.16", features = [
 axum = "0.8.4"
 
 # Required for `turbomcp` example
-turbomcp = { version = "3.0.0", features = ["http", "websocket"] }
-turbomcp-server = { version = "3.0.0", features = ["http", "websocket"] }
-turbomcp-client = { version = "3.0.0", features = ["http"] }
-turbomcp-protocol = "3.0.0"
-turbomcp-transport = { version = "3.0.0", features = ["http"] }
+turbomcp = { version = "3.0.11", features = ["http", "websocket"] }
+turbomcp-server = { version = "3.0.11", features = ["http", "websocket", "channel"] }
+turbomcp-client = { version = "3.0.11", features = ["http"] }
+turbomcp-protocol = "3.0.11"
+turbomcp-transport = { version = "3.0.11", features = ["http"] }
 
 # required for otel
 opentelemetry = "0.30.0"
